feat(fuzzer): add an optional timeout argument
LibAFL uses a default 5-second timeout that is suitable for software
fuzzing, but large-scale RTL models run very slowly. For example,
XiangShan currently runs for 6 seconds for 10K cycles.

This commit increases the default timeout to 10 seconds and allows
an optional command line argument from the user to set the timeout.